+++ pages/profile.tsx
@@ -2,6 +2,16 @@ import HobbyCardEditor from "@/components/hobbyCardEditor";
 import { Button, Col, Container, Row, Form, Stack, Alert, Navbar } from "react-bootstrap";
 
 export default function Profile() {
+
+    fetch("/api/search", {
+        method: "POST",
+        headers: {'Content-Type': 'application/json'},
+        body: JSON.stringify({data: "Some Data"})
+      })
+        .then((res) => res.json())
+        .then((data) => {
+          console.log(data);
+      });
     
     return(
         <>
